refactor(analysis-types): move entropy serde tests
